src/requirements.txt
fastapi==0.111.1
uvicorn[standard]==0.30.1
python-docx==1.1.0
pdfplumber==0.11.4
python-multipart==0.0.9
pytesseract==0.3.10
Pillow==10.4.0
pydantic==1.10.15
requests==2.32.3
pytest==8.3.2
google-cloud-firestore==2.16.1
fugashi[unidic-lite]==1.3.0
